Refactor cache into a util module.
