+++ lib/install.js
@@ -7,12 +7,12 @@ exports.usage = 'Attempts to install pre-built binary for module';
 var fs = require('fs');
 var path = require('path');
 var zlib = require('zlib');
-var os = require('os');
 var log = require('npmlog');
 var existsAsync = fs.exists || path.exists;
 var versioning = require('./util/versioning.js');
 var testbinary = require('./testbinary.js');
 var clean = require('./clean.js');
+var cache = require('./util/cache');
 
 var npgVersion = 'unknown';
 try {
@@ -20,43 +20,6 @@ try {
     var ownPackageJSON = fs.readFileSync(path.join(__dirname, '..', 'package.json'), 'utf8');
     npgVersion = JSON.parse(ownPackageJSON).version;
 } catch (e) {}
-
-function unlink_and_ignore(path) {
-    fs.unlink(path, function (err) {
-        if(err) {
-            if(err.code !== 'ENOENT') {
-                log.verbose('cache', 'removing cache failed', err);
-            }
-            return;
-        }
-        log.verbose('cache', 'cache removed');
-    });
-}
-
-function retrieve_from_cache(opts,callback) {
-    var cache_rs;
-    try {
-        fs.mkdirSync(opts.cache_dir);
-    } catch (e) {
-        if(e.code !== 'EEXIST') {
-            log.warn('cache', 'could not create cache dir ' + e.message);
-            return callback(e);
-        }
-    }
-    cache_rs = fs.createReadStream(opts.cache_path);
-    cache_rs.on('error', function(err) {
-        if(err.code === 'ENOENT') {
-            log.verbose('cache', 'cache not found');
-        } else {
-            log.warn('cache', 'error while reading cache', err);
-        }
-        cache_rs.removeAllListeners();
-        return callback(err);
-    });
-    cache_rs.on('open', function readFromCache() {
-        callback(null, cache_rs);
-    });
-}
 
 function download(uri,opts,callback) {
     log.http('GET', uri);
@@ -109,14 +72,7 @@ function download(uri,opts,callback) {
                 err.statusCode = res.statusCode;
                 return callback(err);
             }
-            if(opts.cache_path) {
-                var cache_ws = fs.createWriteStream(opts.cache_path);
-                cache_ws.on('error', function(err) {
-                    log.warn('cache', 'could not write to cache ' + err.message);
-                    unlink_and_ignore(opts.cache_path);
-                });
-                req.pipe(cache_ws);
-            }
+            cache.write_to_cache(req, opts);
             callback(null,req);
         });
     }
@@ -159,14 +115,13 @@ function extract(source,to,callback) {
 
 function place_binary(from,to,opts,callback) {
     if(!opts.ignore_cache) {
-        retrieve_from_cache(opts,function(err,stream) {
+        cache.retrieve_from_cache(opts,function(err,stream) {
             if(err) {
-                unlink_and_ignore(opts.cache_path);
                 return do_download();
             }
             extract(stream,to,function(err) {
                 if(err) {
-                    unlink_and_ignore(opts.cache_path);
+                    cache.unlink_and_ignore(opts.cache_path);
                     return do_download();
                 }
                 log.verbose('cache', 'binding extracted from cache');
@@ -237,12 +192,7 @@ function install(gyp, argv, callback) {
         var binary_module = path.join(to,opts.module_name + '.node');
 
         opts.ignore_cache = gyp.opts['ignore-cache'];
-        try {
-            opts.cache_dir = process.env.NODE_PRE_GYP_CACHE || process.env.npm_config_node_pre_gyp_cache || path.join(os.homedir(), '.node-pre-gyp');
-            opts.cache_path = path.join(opts.cache_dir,versioning.eval_template(versioning.default_package_name, opts));
-        } catch(e) {
-            opts.ignore_cache = true;
-        }
+        cache.configure_cache_path(opts);
 
         existsAsync(binary_module,function(found) {
             if (found && !update_binary) {
